+++ examples/together-ai/vector-index.ts
@@ -2,34 +2,37 @@ import fs from "node:fs/promises";
 
 import {
   Document,
-  OpenAI,
-  serviceContextFromDefaults,
+  TogetherEmbedding,
+  TogetherLLM,
   VectorStoreIndex,
+  serviceContextFromDefaults,
 } from "llamaindex";
 
 async function main() {
-  // Load essay from abramov.txt in Node
-  const path = "node_modules/llamaindex/examples/abramov.txt";
+  const apiKey = process.env.TOGETHER_API_KEY;
+  if (!apiKey) {
+    throw new Error("Missing TOGETHER_API_KEY");
+  }
+  const path = require.resolve("llamaindex/examples/abramov.txt");
   const essay = await fs.readFile(path, "utf-8");
 
-  // Create Document object with essay
   const document = new Document({ text: essay, id_: path });
 
-  // Split text and create embeddings. Store them in a VectorStoreIndex
   const serviceContext = serviceContextFromDefaults({
-    llm: new OpenAI({ model: "gpt-4" }),
+    llm: new TogetherLLM({ model: "mistralai/Mixtral-8x7B-Instruct-v0.1" }),
+    embedModel: new TogetherEmbedding(),
   });
+
   const index = await VectorStoreIndex.fromDocuments([document], {
     serviceContext,
   });
 
-  // Query the index
   const queryEngine = index.asQueryEngine();
+
   const response = await queryEngine.query({
     query: "What did the author do in college?",
   });
 
-  // Output response
   console.log(response.toString());
 }